+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.27.0 - 2020-01-20
+- **BREAKING** Rename: all `rawCheck` fields to `check`
+- Update: allow use `RedeemCheckTxData` without `proof` and `privateKey` it will lead to empty proof
+
 ## 0.26.1 - 2020-01-20
 - Add: option for `prepareLink` to allow setting custom hostname